refactor: promote exact award identity
