+++ packages/sbir-analytics/sbir_analytics/assets/phase_iii_negative_controls/nih_reporter.py
@@ -11,7 +11,7 @@ from typing import Any
 
 import pandas as pd
 
-from .identity import IdentityRecoveryError
+from sbir_etl.identity.exact_awards import IdentityRecoveryError
 from .source_keys import NIH_CORE_PROJECT_ADAPTER, NIH_PROJECT_ADAPTER
