+++ sbir_etl/identity/__init__.py
@@ -12,6 +12,14 @@ from .company_names import (
     rapidfuzz_token_set_100,
     rapidfuzz_token_sort_100,
 )
+from .exact_awards import (
+    EXACT_AWARD_IDENTITY_VERSION,
+    ExactAwardIdentityProfile,
+    IdentityRecoveryError,
+    RecoveryStatus,
+    reconcile_award_identity_attempts,
+    resolve_award_identities,
+)
 from .geography import (
     US_JURISDICTION_NAMES_V1,
     US_JURISDICTION_VARIATIONS_V1,
@@ -24,6 +32,7 @@ from .geography import (
 
 __all__ = [
     "ENHANCED_ABBREVIATIONS",
+    "EXACT_AWARD_IDENTITY_VERSION",
     "SUFFIX_TOKENS",
     "US_JURISDICTION_NAMES_V1",
     "US_JURISDICTION_VARIATIONS_V1",
@@ -31,6 +40,9 @@ __all__ = [
     "VALID_US_JURISDICTION_CODES_V1",
     "CompanyNameMetric",
     "CompanyNameProfile",
+    "ExactAwardIdentityProfile",
+    "IdentityRecoveryError",
+    "RecoveryStatus",
     "company_name_similarity",
     "normalize_company_name",
     "normalize_us_jurisdiction",
@@ -38,5 +50,7 @@ __all__ = [
     "rapidfuzz_ratio_100",
     "rapidfuzz_token_set_100",
     "rapidfuzz_token_sort_100",
+    "reconcile_award_identity_attempts",
+    "resolve_award_identities",
     "us_jurisdiction_name",
 ]
