+++ tests/unit/scripts/test_identity_boundaries.py
@@ -91,5 +91,19 @@ def test_small_state_fixture_is_not_mistaken_for_an_implementation(tmp_path: Pat
     assert boundaries.scan_file(path, repository_root=tmp_path) == []
 
 
+def test_duplicate_exact_award_resolver_is_rejected(tmp_path: Path) -> None:
+    path = _write(
+        tmp_path,
+        "packages/example.py",
+        "def resolve_award_identities(sbir, official):\n"
+        "    return sbir.merge(official, on='award_id')\n",
+    )
+
+    violations = boundaries.scan_file(path, repository_root=tmp_path)
+
+    assert len(violations) == 1
+    assert "exact award-key resolver" in violations[0].message
+
+
 def test_current_repository_obeys_identity_boundaries() -> None:
     assert boundaries.scan_repository() == []
